Snapshot/live symbol request data
Part of issue #2.
Will properly request live symbol data on inspection & clicking on
subelements of a list.
Currently a problem with symbolId being a Proxy object, which causes an
Immutable error as well as nothing happening on list element click.
Likely a Promise issue? Will ask @ryanholmdahl  to take a look.