pass all primitives by reference
All primitives _which should be openable in a viewer_ are now passed by reference. This means fields where whose values are created by us (and are not Python objects created in the user's program) and values that should be "shown" rather than treated as objects (like the string names of arguments to a function) are not passed by reference. Ultimately, this should leave us with basically no "surprise" special cases, as we had before in `SequenceViewer`.
Resolves #22 .