The death of the live viewer
I have embarked on a glorious crusade, purging this program of the wretched distinction between snapshot and live viewer. There is now only the snapshot viewer, a glorious hybrid of entities which when separated were of unquesitonable evil. Where once we had a viewer which captured historical state but was not interactive, and another which was interactive but reflected only the end of the program, by act of divine providence we find only one which both captures historical state and allows one to explore it interactively. We shall not mourn the loss of the live viewer, but shall carry it onward as a reminder of our past sins and of the redemption we must humbly beg of our Lord.