Band-aid for immutability issues in setState()
When state.set() is called in a the callback passed to setState(), an error is thrown, because state is not an Immutable anymore. It seems there are some issues with wrapping the entire state object as an Immutable. For now, I just re-wrap the state object in the callback, but we should either determine how to properly use Immutable here or otherwise abandon it.